server/venv_requirements.txt
alembic==1.8.1
autopep8==1.7.0
bcrypt==4.0.0
bson==0.5.10
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
Mako==1.2.2
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
pycodestyle==2.9.1
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
rptree==0.1.1
six==1.16.0
SQLAlchemy==1.4.40
toml==0.10.2
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
gunicorn==20.1.0
